requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
rank-bm25==0.2.2
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
numpy==1.26.4
tenacity==9.0.0
